requirements.txt
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
dbr==9.0
numpy==1.22.3
opencv-python==4.5.5.64
pdf2image==1.16.0
pdfminer.six==20220506
Pillow==9.1.0
pycparser==2.21
pyzbar==0.1.9